pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "omegalax"
version = "0.1.0"
description = "JAX-based VLM training codebase scaffold."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "absl-py",
  "etils",
  "flax>0.12.0",
  "jax[cuda12]",
  "jaxtyping",
  "optax",
  "safetensors",
  "transformers",
  "huggingface-hub",
]

[project.optional-dependencies]
torch-tests = [
  "torch==2.10.* ; platform_system == \"Linux\"",
]

[tool.hatch.build.targets.wheel]
packages = ["omegalax"]
